+++ source/blender/src/drawmesh.c
@@ -142,7 +142,7 @@ static int draw_tfaces3D__setHiddenOpts(void *userData, int index)
 {
 	struct { Mesh *me; EdgeHash *eh; } *data = userData;
 	MEdge *med = &data->me->medge[index];
-	unsigned long flags = (long) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
+	uintptr_t flags = (intptr_t) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
 
 	if((G.f & G_DRAWSEAMS) && (med->flag&ME_SEAM)) {
 		return 0;
@@ -160,7 +160,7 @@ static int draw_tfaces3D__setSeamOpts(void *userData, int index)
 {
 	struct { Mesh *me; EdgeHash *eh; } *data = userData;
 	MEdge *med = &data->me->medge[index];
-	unsigned long flags = (long) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
+	uintptr_t flags = (intptr_t) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
 
 	if (med->flag&ME_SEAM) {
 		if (G.f&G_HIDDENEDGES) {
@@ -176,7 +176,7 @@ static int draw_tfaces3D__setSelectOpts(void *userData, int index)
 {
 	struct { Mesh *me; EdgeHash *eh; } *data = userData;
 	MEdge *med = &data->me->medge[index];
-	unsigned long flags = (long) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
+	uintptr_t flags = (intptr_t) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
 
 	return flags & eEdge_Select;
 }
@@ -184,7 +184,7 @@ static int draw_tfaces3D__setActiveOpts(void *userData, int index)
 {
 	struct { Mesh *me; EdgeHash *eh; } *data = userData;
 	MEdge *med = &data->me->medge[index];
-	unsigned long flags = (long) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
+	uintptr_t flags = (intptr_t) BLI_edgehash_lookup(data->eh, med->v1, med->v2);
 
 	if (flags & eEdge_Select) {
 		return 1;
